Move settings package
