+++ format/it/it.go
@@ -4,7 +4,7 @@ package it
 import (
 	"github.com/gotracker/playback"
 	"github.com/gotracker/playback/format/it/load"
-	"github.com/gotracker/playback/format/settings"
+	"github.com/gotracker/playback/settings"
 )
 
 type format struct{}
