+++ format/it/load/formatutils.go
@@ -3,7 +3,7 @@ package load
 import (
 	"github.com/gotracker/playback/format/it/layout"
 	"github.com/gotracker/playback/format/it/playback"
-	"github.com/gotracker/playback/format/settings"
+	"github.com/gotracker/playback/settings"
 )
 
 type readerFunc func(filename string, s *settings.Settings) (*layout.Song, error)
