+++ format/it/load/load.go
@@ -2,7 +2,7 @@ package load
 
 import (
 	"github.com/gotracker/playback"
-	"github.com/gotracker/playback/format/settings"
+	"github.com/gotracker/playback/settings"
 )
 
 // IT loads an IT file
