+++ format/s3m/load/formatutils.go
@@ -3,7 +3,7 @@ package load
 import (
 	"github.com/gotracker/playback/format/s3m/layout"
 	"github.com/gotracker/playback/format/s3m/playback"
-	"github.com/gotracker/playback/format/settings"
+	"github.com/gotracker/playback/settings"
 )
 
 type readerFunc func(filename string, s *settings.Settings) (*layout.Song, error)
