+++ format/s3m/load/load.go
@@ -4,7 +4,7 @@ import (
 	"github.com/gotracker/playback"
 	"github.com/gotracker/playback/format/s3m/layout"
 	"github.com/gotracker/playback/format/s3m/load/modconv"
-	"github.com/gotracker/playback/format/settings"
+	"github.com/gotracker/playback/settings"
 	formatutil "github.com/gotracker/playback/util"
 )
