+++ format/xm/playback/playback_test.go
@@ -7,9 +7,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/gotracker/playback/format/settings"
 	"github.com/gotracker/playback/format/xm"
 	"github.com/gotracker/playback/player/feature"
+	"github.com/gotracker/playback/settings"
 	"github.com/gotracker/playback/song"
 )
