+++ format/xm/xm.go
@@ -3,8 +3,8 @@ package xm
 
 import (
 	"github.com/gotracker/playback"
-	"github.com/gotracker/playback/format/settings"
 	"github.com/gotracker/playback/format/xm/load"
+	"github.com/gotracker/playback/settings"
 )
 
 type format struct{}
